requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.2
django-dotenv==1.4.2
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
pillow==10.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
zope.interface==6.2
